Update MariaDB to 10.5
This bumps the MariaDB PPA from 10.2 to 10.5.

Note that this will *not* actually upgrade the version of MariaDB
already installed on a server.

Upgrading is a manual process that involves:

1. stopping the service
2. removing the installed package
3. re-installing the package

The full process is detailed at https://mariadb.com/kb/en/upgrading-from-mariadb-104-to-mariadb-105/

Though it's not required, backing up your database before the upgrade is
recommended.